Two issues are left cuz IDK how to fix them.

+++ src/Core/IndexBuffer.cpp
@@ -2,7 +2,7 @@
 
 namespace Core {
 IndexBuffer::IndexBuffer(const std::vector<unsigned int> &indices)
-    : m_Count(indices.size()) {
+    : m_Count(static_cast<unsigned int>(indices.size())) {
     glGenBuffers(1, &m_BufferId);
     glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, m_BufferId);
     glBufferData(GL_ELEMENT_ARRAY_BUFFER,
